SUCCESS FINISHED

+++ src/main/java/com/example/taskitmed/repository/IdentifierRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author: JADANOV_DILMUROD
  * @date:28.05.2022
  */
-public interface IdentifierRepository extends JpaRepository<Identifier, Long> {
+public interface IdentifierRepository extends JpaRepository<Identifier, Integer> {
 }
